feature: saves (#43)
* build: add RocksDB dependency

* refactor(gameplay): organize client and server code into subdirectories

* refactor(client): use random_device for world seed

Replace `ChunkGenerator::seed()` with `std::random_device` in `ClientWorld::init` to remove dependency on server-specific chunk generator code.

* refactor(server): move heightmap ownership to chunk generator

Heightmap is only used during generation, so it doesn't belong in
chunk data. Add storage data helpers for future persistence.

* feat(server): add RocksDB chunk persistence

Add a ChunkStorage class backed by RocksDB with protobuf serialization for saving chunk data. Chunks are now loaded from storage before generation, and generated or removed chunks are persisted individually or in batches. Introduce save_all_chunks on stop and configurable zstd compression level.

* feat: add world name customization and enhanced chunk saving

Added --world CLI option and host game UI field for world name. Passed
world name through server startup and chunk storage path. Added periodic
auto-save for server chunks. Fixed chunk save path concatenation. Guarded
against empty chunk serialization and unsupported chunk versions.
Null-checked client session on close.

* fix(chunk-system): stop generation before saving chunks on shutdown

Stop generation threads and pool before saving all chunks to avoid
writes during shutdown. Add size() to ChunkStorage, replace unused
methods, and add profiling scopes for chunk system operations.

* fix(chunk-storage): improve save/load error handling

Return false from save_batch when any chunk fails to serialize, and
fail deserialization if block count is invalid or biome lookup throws.
Only log save success when the operation actually succeeds.

* refactor(server): make block access thread-safe

Return block vectors by value under a shared mutex, add world-coordinate get/set block accessors, and load neighbors from storage when available instead of generating them. Save removed chunks synchronously.

* feat(server): persist world seed and tick metadata

Add world metadata.json to store seed, game ticks, and day ticks.
Initialize the chunk generator from saved or provided seed when starting a world.
Add JSON serialization utilities and update world scene startup to pass seeds through the server.
Update chunk storage key format.

* feat(tools): add atomic file replacement for JSON saves

Write JSON to a temp file in the same directory before replacing the target. This prevents partial writes from corrupting files. Added `atomic_replace` helper with Win32 and POSIX implementations.

* feat(tools): add cross-platform StandardPaths helper

Add header-only StandardPaths utility for Windows and Linux
providing standard config, data, cache, runtime, and user
directories per XDG and Windows known folders.

* refactor(server): centralize storage in WorldStorage

Add WorldStorage to manage shared RocksDB instance, refactor ChunkStorage to use it, introduce EntityStorage for entity persistence, and rename net_utils to proto_utils.

* feat(server): implement entity persistence

Load entities from storage on init, save on world save and unload,
and unload entities when their chunk is no longer loaded. Also
persist the next entity ID in world metadata.

* refactor(server): centralize entity count and add stop method

Add ServerEntityManager::stop() to save all entities on world shutdown.
Move m_entity_sum increment into the common entity insertion helper so
the count is updated in one place. Add lifecycle log messages.

* fix(server): increment entity count only on successful emplace

* feat(server): add chunk-based dormant entity activation

Entities are now loaded as dormant data grouped by chunk, activated only when their chunk finishes generating, and returned to dormant storage when the chunk becomes inactive. This prevents entities in unloaded chunks from being ticked and improves unload/destroy handling.

* refactor(world-storage): store metadata in RocksDB instead of file

* refactor(server): use common WorldStorage version constant

* feat: add world name validation

- Add Cubed::Tools::is_valid_world_name to enforce portable, safe save file names.
- Validate world names in the host game UI and server world initialization.
- Prevent path traversal by ensuring the save path remains within ./saves.
- Add localized error message for invalid world names.

* fix(world-storage): handle missing metadata without error

- Treat RocksDB NotFound as expected absence and return nullopt directly.
- Only log an error for actual read failures.

+++ include/Cubed/gameplay/game_time.hpp
@@ -3,11 +3,12 @@
 // Prevent unsigned underflow issues in subtraction
 #include "Cubed/tools/cubed_assert.hpp"
 
+#include <cstdint>
 #include <functional>
 #include <utility>
 namespace Cubed {
 
-using TickType = long long;
+using TickType = int64_t;
 using TimeType = float;
 
 constexpr int DEFAULT_PER_TICK_TIME = 50;
